Fire delay issue - fix
Fixed the fire delay issue and introduced events and the appropriate methods to facilitate the change of other player stats. Added a sprite for the cursor.

+++ Assets/Scripts/Items/ItemInformationSO.cs
@@ -12,7 +12,7 @@ namespace Items
         public float damageDelta;
         public float damageMultiplierDelta;
         public float healthDelta;
-        public float speedDelta;
+        public int speedDelta;
         public float bulletLifetimeDelta;
         public float fireDelayDelta;
         public float fireRateMultiplierDelta;
